refactor(client): split recovery wizard into client page module
Move IndexedDB recovery UI into recovery-client with reducer-driven state so
the route exports metadata from a thin server page.
